backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
dashscope==1.14.0
pydantic==2.5.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
